Rename aiohabit project to mrack
