+++ scripts/mrack
@@ -1,3 +1,6 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
 # Copyright 2020 Red Hat Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,5 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from mrack import run
 
-"""aiohabit actions."""
+if __name__ == "__main__":
+    run.run()
