+++ src/mrack/__init__.py
@@ -1,4 +1,4 @@
-"""aiohabit library."""
+"""mrack library."""
 import logging
 
 logger = logging.getLogger(__name__)
@@ -8,7 +8,7 @@ logger.setLevel(logging.DEBUG)
 file_formatter = logging.Formatter("%(asctime)s %(name)s %(message)s")
 console_formatter = logging.Formatter("%(name)s %(message)s")
 
-file_handler = logging.FileHandler("aiohabit.log")
+file_handler = logging.FileHandler("mrack.log")
 file_handler.setFormatter(file_formatter)
 
 console_handler = logging.StreamHandler()
