+++ src/mrack/dbdrivers/__init__.py
@@ -1,6 +1,3 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
 # Copyright 2020 Red Hat Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,7 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from aiohabit import run
 
-if __name__ == "__main__":
-    run.run()
+"""mrack database drivers.
+
+A database driver persists and loads results of actions so that it can be used in
+consecutive actions.
+"""
