+++ src/mrack/dbdrivers/file.py
@@ -15,8 +15,8 @@
 """File database driver module."""
 
 from os import path
-from aiohabit.host import host_from_json
-from aiohabit.utils import save_to_json, load_json
+from mrack.host import host_from_json
+from mrack.utils import save_to_json, load_json
 
 HOSTS_KEY = "hosts"
