+++ src/mrack/transformers/aws.py
@@ -14,8 +14,8 @@
 
 """AWS transformer module."""
 
-from aiohabit.transformers.transformer import Transformer
-from aiohabit.utils import get_config_value, print_obj
+from mrack.transformers.transformer import Transformer
+from mrack.utils import get_config_value, print_obj
 
 CONFIG_KEY = "aws"
